requirements.txt
click==6.7
Flask==0.12.2
Flask-HTTPAuth==3.2.3
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.3.2
SQLAlchemy==1.1.11
Werkzeug==0.12.2
click==6.7
requests==2.13.0
plaid-python>=2.0,<2.1
